engine:Add quota validation for crateAllSnapshot
In the current patch, I added support for validating quota when
creating snapshot
The validation assembled from several phasas, and count on a delta
memory map mechanism, described in the quota manager.

The validation start with the method validateQuota, which
update the delta map with the storage resources demended for the
operation about to be executed.
After the operation finishes, and the VM is running,
The delta size added before is being removed.

Change-Id: I72714e6aa0e5ff034c1192297743dbe79c0d351b