core: VDSGroup implements BusinessEntity
no bz

The business entities have common procedures- getId and setId.
But as they don't have a common interface, each business entity has
getId and setId methods of its own. Even the names of these methods in
different business entities are not the same
(some have getId other getID and ect.)

In this patch I added to the VDSGroup implemented interfaces-
BusinessEntity inteface.
In VDSGroup- refactoring getID() to getId() and setID(...) to setId(...).

Change-Id: Idd48e8eb3d63bb5e0fa334c625e2e342dc598fca