core: Added correlation ID to VdcReturnValueBase
A new field, correlationId was added to VdcReturnValueBase. The
purpose of the field is to identify a cross system action, start by
the clients thru the backend till VDSM.

The value of the correlation id will contain the correlation id either
as provided by the client or as generated by the ovirt-engine and used
to log the action.

Change-Id: If360d44f9713aa3fbad49acfa47cad1f4f3a445b