engine: Add quota validation for Add Disk to VM
In the next patch, I added support for validating quota when adding new
disk.
The validation assembled from several phasas, and count on a delta
memory map mechanism, described in the quota manager.

The validation start with the method validateQuota, which then update
the delta map with the size requested to consume.
After the operation finished, we remove the delta that was saved before
from the delta map.

Change-Id: I35f205da64ef101b4e38cd7a99ab88142a3e277b