engine:Add quota validation for RunVmCommand
In the next patch, I added support for validating quota when running VM
The validation assembled from several phasas, and count on a delta
memory map mechanism, described in the quota manager.

The validation start with the method validateQuota, which
update the delta map with the vdsGroup resources demended for the
operation about to be executed.
After the operation finishes, and the VM is running,
The delta size added before is being removed.

Change-Id: I476a53ef9a479bb7a26fdec45abd492268c2d3d6