engine: Add quota validation support commandBase
In the next patch, I added support in command base for quota
consumption.

Quota consumption assembled with several phasas, and count on a delta
memory map mechanism, described in the quota manager.

The validation start with the method validateQuota, which update
the delta map with the size supposed to be consume in the process about
to be executed.
After the operation finishes, whether it succeeded or not, the delta
that was saved before is removed from the memory map by calling remove
quota left over method.

Change-Id: Ie1f25eb05266ba7e0b777fb7ad3d138a99fd1477