userportal: extended->templates->network interfaces->first three tabs
Implementation of the first three tabs (general, network interfaces
and virtual disks)

This patch includes the extraction of the respective logic from
WebAdmin and also the refactoring of the WebAdmin to use the extracted
logic.

This patch introduces the UserPortalSearchableDetailModelProvider
which is a base class for the providers for searchable detail models.

This patch does not include the handling of the action buttons on the
subtabs.

Change-Id: I0410519c2377aac9b1aa490f20b734074e3b9fdd