core: vm_device upgrade script patch
This is a fix to a bug in the upgrade script that added disks to vm_device with
the image_id rather than image_group_id.

Also fixed monitor creation in vm_device for Desktop VMs only

Patches 03_01_0220 and 03_01_0370 execution was commented since those patches and the fix
builds vm_device table from scratch, so , only the latest upgrade should be executed.

Change-Id: Ic3a5e053f977c74dcf24dc3feb1467960825c7fc