core: Adding new query
Adding the query GetVmConfigurationBySnapshot.
The snapshot will be obtained based on the snapshotID from
the DB, and its configuration will be parsed and returned to
the user.
As this is a user query, a new version of "get" method at
SnapshotDAO was added, accepting also userID and the filter
flag (The permission is checked on the VM containing the
snapshot).
SnapshotDaoTest was modified accordingly as well

Change-Id: I3155134ff9ad89bb50ec4baa8fd393844f80bb32
Acked-By: Mike Kolesnik