core: Change images_storage_domain_view to left join disks
This is required for supporting preview & restore from snapshot to
snapshot.
View was also fixed so that the switch cases don't return false values
in the case that the fields are null (which can happen because of left
join).

Needed to explicitly ignore the "Blank template image" in the view,
since it is a fake entry and should not be returned.

Change-Id: I54a5022784116acbd3e1daac41657cc096d62ef3