userportal: extended -> templates -> events
Implementation of events subtab of the templates side-tab of the
extended tab

This patch also contains the generalization of the table which renders
the AuditLog - the AuditLogModelTable. This is a general table used for
all tabs in Webadmin and UserPortal to draw the Events tabs.

This patch does not include the handling of the action tables

Change-Id: I6c4f5879423ab6d92d5ac6cd2738b9c8d935d72b