userportal: extended tab->templates->permissions
Implementation of the Permissions subtab of the template side tab of
the extended tab.

This patch also includes the generalization of the table which renders
the permission tabs - the PermissionListModelTable.
The default implementation contains only the 'User' and 'Role' columns. It
is than extended by the PermissionWithInheritedPermissionListModelTable
which adds the 'Inherited Permission' column.

This patch does not include the handling of the action buttons.

Change-Id: Ifd97ced9ea4efd6094729556515c9da0fa61f288