core: Filter GetClustersWithPermittedActionQuery
Define the query as a user query so it could be called in the User
Portal.

The query was refactored to use the user stored in the query, which
is initiallized from SessionDataContainer in the ctor instead of
calling it again.

A test was written to formally prove that the behavior was not changed
(except for the definition as a user query).

Change-Id: Ifaba46314597b607163c9bb43dc4e7f1440a7d6c