webadmin: Only partial menu options are enabled on first right-click
https://bugzilla.redhat.com/show_bug.cgi?id=784189

If you do a right click on an unselected table row the opened context menu
is updated according to the previous selection.
That happens because on right click the selection is updated only after
scheduleDeferred and not immediately. When the onConextMenu
(AbstractActionPanel) is called the selection is not updated yet.

Change-Id: I4b8de1438cd3635812d67429bc46606fdd1e8eaa