engine: Refactor permission subject
Adding new class for permission which encapsulate the permission
parameters such as entity id, entity type and action group.
The motiviation to do so, is to add support for validating permssions
on empty action groups.

The next patch replace the use with map to be used with the new
permission subject class

Change-Id: Ie4e126717f848bcee816c2a81fd00f84e1448e1c