core: Fixing vds status when installed without boot.
This patch sets non-responsive status to a host which is being
installed without reboot, so the host will be polled. Hosts
which are rebooting gets a Reboot status until the reboot timeout
will end.

Another issue handled here, is improving the replacement of the reboot
flag, in order to prevent other command parts to be duplicated.

Change-Id: I831659ae09fdd356dd17f185bb28da68066b3e31