core: Add monitoring for Maintenance Host action
The patch adds monitoring for the Maintenance Host command.
The MaintananceNumberOfVdssCommand will not be monitored since is not
granular enough (might contain a hosts from different clusters)).
As part of the Maintenance, VM migrations will be monitored as well.

Since Maintenance might be failed/aborted, marking a running Maintenance
Job as completed is spread also to ActivateVdsCommand and
VdsEventListener where the Job is marked as failed.

Change-Id: Id8c006a4ed26ada45802af735041819b7cdabce2