core: New mechanism in CommandBase (rollback using command)
This mechanism will allow invoking internal commands from
endWithFailure in order to perform rollback at backend.
An example for this is a need to remove a destination image
 after it was successfuly copied (as a result of failure of
another CopyImage).
This mechanism makes sure that the RemoveImage will not end
in endSuccessfuly, even if successful at VDSM, but will end
in endWithFailure.

Change-Id: Ic734d785215de855817acf77c88cba9047b959eb
Acked-By: Mike Kolesnik <mkolesni@redhat.com>