core: Make dependency on log4j explicit
There are several places in the code where we are importing
classes from log4j, but we don't have an explicit dependency on
log4j in the main POM file or in the specific POM files for the
affected components. Right now this is working because Maven is
bringing that transitive dependency (incorrectly, I guess). In
Maven 3 this doesn't work.

This patch adds explict dependecies where needed.

Change-Id: Id2e85a154bf5eb94a6710a17933d6160b12daec4