core: vm_device upgrade script patch
This is the first patch in a patch series for the Stable Device Addresses feature.
The first series (phase 1) was just to change the data model & the way we talk with VDSM
The second series (phase 2) is actually getting the addresses when a VM is in 3.1 cluster and store them in DB
then those addresses are used when a VM starts to run
See full detail on oVirt wiki site :
http://www.ovirt.org/wiki/Features/Design/StableDeviceAddresses

This is a fix to a bug in the upgrade script that added disks to vm_device with
the image_id rather than image_group_id.

Patches 03_01_0220 and 03_01_0370 execution was commented since those patches and the fix
builds vm_device table from scratch, so , only the latest upgrade should be executed.

Change-Id: I4685889e4651a531ba15182aa516221b2f2ef730