core: add FullListVdsCommand support
VDSM added a new list signature that accepts mode and a collection of VMs and returnes the full (mode=full)
device description of each VM in the given collection.
This is used when we are issuing getAllVmStats that returns a hash for each VM , then , we collect all VMs
with dirty hash and sent it at once to VDSM to get the changes and persist them.

Change-Id: Ieda51634469d3925a9afa89970d60a5427144e98