core: move interfaces and disks collections to VMBase
This patch moves the collections of interfaces & images/disks from VM to VmBase since the
same collections exists in VmTemplate and Vmstatic & VMTemplate are inherited from VmBase.
This simplifies the code especialy in VmDeviceUtils where code is now more generic.

However, right now only the collections were moved and the original getter/setter in VM delegates
the code in VmBase.

More work to be done to have all code call VmBase code directly, but this involves lot of modifications is
is out of the scope of this patch set.

Change-Id: I18be17cdcad37d4940b99cbad6dc3abaf04d8595