core: adding service strategy for gluster integration
This commit adds capability to use a strategy on service-related
operations, such as initilizing a host, monitoring it and etc.
The purpose in this commit is to enable integrating different service,
that each one has a different strategy in different scenarios. Each such
strategy will be defined using a method in the service strategy interface.

Change-Id: I5edc263dae23236015260d83a933e8a3a93e4523