webadmin: adding support to downdstream dashboards
no bz

In downstream we have dashboard with iframes and in upstream we have
html page with no iframes. So we have to pass different parameters in order
to get the "dashboard".

Change-Id: Icff512cc7e8b48a4494fd3584081c43abbbeabc1

webadmin: adding dashboards to Report.xml

no bz

Adding dashboards to Report.xml in order not to have the downstream urls
in the webadmin code

Change-Id: I9704b2ec21eb95c9ff8655fe7ddd93d17cefd61f