core: Run upgradeStoragePool on cluster compatibility change
This will cause a pool to upgrade when the compatibility version is
raised appropriately.

The domains are marked as updated in the DB even though it might take
some time untill they are actually updated. This is OK as VDSM will
update them as long as they are accessed with the correct storage format
type.

Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Change-Id: I85201624bbb2f8c41cf3b184b89a8e199ff50e99