core: getting disk of vm and templates differently
This patch includes also minor change of fixing NPE (the setting of the vm variable)
However, main thing to look at is that from some reason there is a duplication of images
collections (moved from VM to VmBase)
One collection is the images collection and the other is the diskList
I have noticed that images is filled correctly for VMs and diskList for images
but, I don't understand the logic and it looks to me like a redundant duplicatiom, however,
refrences to both collections are spread all over the code and I have no time to explore that nor
doing a violant change.
I will be happy to get any comments or suggestions for improvements on that.

Change-Id: I9726492eb401b5c86562757905c2fd84b7c12ffc