userportal: VM view -> console buttons
Implementation of the open console and edit console buttons on the rows
of the VM view on the extended tab.

This patch also includes a extraction of the logic for opening a
console from MainTabBasicListItemPresenterWidget to injectable class
ConsoleManager.

This patch also changes the ConsolePopupView to not use editor driver,
as on cancel the model should not be changed.

Change-Id: Id7856b8eeb2537431abab1eef6076cb5b8faf3cf