core: Restore missing image as ILLEGAL
If image is missing, this means it got deleted somehow, so restore it as
ILLEGAL image so that the user can know that there was a disk in that
snapshot.
Handling of ILLEGAL image was added to RestoreAllSnapshotsCommand as
well since these images need not be restored to, and the can do action
should ignore checking for missing images since they can be present.

Change-Id: I77f7ebdaa1f558351f657245b6dd395e1adad976