webadmin: RunOnce dialog - Domain is always visible
https://bugzilla.redhat.com/show_bug.cgi?id=784852

In "RunOnce" dialog the entire "Windows Sysprep" section should be
displayed only if the "Attach Floppy" check-box is checked and the
selected floppy in the drop-down is "[sysprep]".
The problem was that on the first run of the vm, the "Windows Sysprep"
section was displayed even if the "Attach Floppy" was unchecked.

Change-Id: I67b6e9bc511dc11fe211717661b55bc3bfa2bc15