userportal: Extended tab->Vm->Subtabs for pools
On the VM side tab of the extended main tab in the main grid both the
VMs and Pools are shown.

The General, Network Interfaces and Virtual disks have different models
for Pools and VMs, so needed to have different model providers and
view-presenter pairs.

It was done by binding all the subtabs (for both VMs and Pools) to the
same ExtendedVmSubTabPanelPresenter and using the ModelBoundTabData to
display only the ones which are relevant for the current
selection. To be able to do it, it was needed to move the TabFactory
from webadmin to gwt-common.

The problem with this approach is, that there will be two
subtabs with the same label (one for Pools, one for VMs). As the
AbstractSideTabWithDetailsPresenter.getSubTabRequest calculates the
request token according to the title of the tab, it was needed to
differentiate between this two cases. To do this, the logic of
calculating the request token has been extracted to standalone method
and overridedden in the SideTabExtendedVirtualMachinePresenter. This
finds out, if the specific token belongs to a pool specific model and if
so, adds the '-pool' suffix to it.

Change-Id: Ic3fb153434fdd0cf83bbd146fdf21e8b4b3d38e5