core: Monitor any actions if executed with a correlation ID
Each action which is executed with a correlation ID should be monitored,
even if the action is configured internally not to monitored. For example,
cluster creation will not be monitored when is executed explicitly, however
as part of the 'Configure local storage' it will be monitored upon
providing a correlation ID by the client.

Change-Id: I27b33897ecdbeb4a7cfc4049aa3dfd516be1c442
Signed-off-by: Moti Asayag <masayag@redhat.com>