webadmin: adjust sub menu popup location to fit screen + RTL popup support
no bz

If the subMenu popup doesn't fit the screen it will be opened in another
position if poosible, or with a scroll bar.

This feature is a workaround to the problem introduced in
http://gerrit.ovirt.org/#change,2489
It should have been implemented in the gwt MenuBar. But it wasn't.
See-
http://code.google.com/p/google-web-toolkit/issues/detail?id=407
http://code.google.com/p/google-web-toolkit/issues/detail?id=3924

In order to implement this workaround I had to use gwt MenuBar private
method. I did it with the help of JSNI.

Change-Id: Ide1b780d9db245f61aa7332a50cf56b3a4761065