userportal: OS type for pool was incorrect
The vm_pools entity does not have a field which defines the OS type
associated with it. To find out what is the specific OS type, an async
request is performed. The problem is this scenario:

1: the pool is returned from the server and the specific
   UserPortalItemModel is filled from it in PoolItemBehavior
   (without the OS type)
2: an async query is fired to the server to find out the OS type
   defined for this pool (from PoolItemBehavior)
3: the pool is shown on the GUI (without the OS type which falls back
   to defaul which is 'Unassigned')
4: the result from the server comes back with the specific OS type and
   this is set to the UserPortalItemModel (which is not reflected
   on the gui)

In this scenario the correct pool OS type is never shown on the GUI.

The fix was to cache the OS type in PoolItemBehavior given from the
async query and if an OS type for the given pool is already cached, it
is set to the specific UserPortalItemModel. In this case, at first the
default 'Unassigned' is shown on the screen, and after the next refresh
the correct OS type is shown.

Change-Id: I336b195adf0b1a72360410118a6b54b674138570