engine: Add test for search query
Adding a test for search query.
The search for now tests the get all queries by checking if the view,
key and the order by, are part of the generated sql.

The tests which were added are testing storage pool, VM, VDS, disk image
and cluster.

Take in notice one test is ignored since searching for multiple
Datacenters is not implemented in the search backend.
So the patch which will fix that should also change the test from
ignored.

Change-Id: I07cc6795008b39281567643697e30b6710f6f457