userportal: Implementation of Network Interfaces popups
The implementation of the Network Interfaces popups for both
VMs and Templates.
The implementation of the widget(NetworkInterfacePopupWidget) is the
same for both.
The difference between the view for VM and for Template is that the
template one has no MAC address configuration. This is solved by hiding
this widgets from the screen, when the getMAC.getIsAvailable() returns
false on the VmInterfaceModel (and set this property in the
TemplateInterfaceListModel).

This patch also includes the following small fixes:

  - UserPortalSearchableDetailModelProvider needs to implement the
    SearchableDetailModelProvider, but it implemented the
    DetailModelProvider (othervise the setSelectedItem will not be
    called on it after the selection is changed in the table)

  - to the TemplateInterfaceListModel.remove() added
    model.setHashName("remove_network_interface_tmps"); which enables
    the RemoveConfirmationPopupView to show the correct message

Change-Id: I27941700901974839175d43a2f4e282cf334613e