core: Fix illegal images handling
Illegal images should be deleted from DB each time we switch active
snapshot since they are not treated by the images command (they might
not be a part of the target snapshot images).

Additionally the created illegal images need to be linked to the active
snapshot which is determined when switching snapshots.

Change-Id: I8defc2660cf9404a088050dd10df8632d79f1de5