webadmin: can't copy from textbox in Host - General subtab (#805173)
https://bugzilla.redhat.com/805173

The problem was, that when the user selected something in the general
subtabs values, after the next refresh, this selection disappeared.
The problem with this was that it was highly problematic to copy some
values from the general subtab which is a common operation.

The behaviour implemented now is the following:
- when the user creates some selection, it will not be lost after new
  data arrives
- when the user right-clicks to a value in the general subtab, the
  whole value gets selected and a browser right-click menu will pop up
  (this selection will not be lost after new data arrives)

To achive this, the following changes had to be done:
- TextBoxLabel recreates the selection in setText
- AbstractFormPanel rememberes a specific instance of the TextBoxLabel
  for every ValueLabel so the listeners will not get lost after refresh

Change-Id: Ie80056c15fbd524dd5e424e246828db91af62114