userportal,webadmin: After logout only a blank screen is shown (#782382)
https://bugzilla.redhat.com/782382

When the user pressed the 'Sign Out' button, only the blank screen was
shown.

This happend because when the login section has been changed to main
section, the hide() has not been called on the LoginPopupView.

This regression has been introduced by the
135e1d00019cdae4dd100a53f84c4d61f2bb3f0c commit, which removed the
setAutoHideOnNavigationEventEnabled(true); from the AbstractPopupView's
initWidget() method, as it made no sense for the rest of the
application, where the popups are managed by UiCommon models.

This patch introduces the class AbstractLoginPopupView which is a base
class for LoginPopupView for both userportal and webadmin and sets the
setAutoHideOnNavigationEventEnabled(true).

This patch also includes the adjustment of ReportInit to be possible to
call it more than one times. The init is called only after login, and
than remembers it's state. This state has to be reseted before every
init() calling. This reset has been added to the initState() method on
the ReportInit class and is invoked as a first statement of the init()
method.

This patch also includes the implementation of the logout for the
userportal-gwtp as this part missed.
It includes:
- calling the Frontend.LogoffAsync
- clearing of the UserPortalLoginModel's content
- listening on the LoginChangeEvent in UserPortalDataBoundModelProvider
  and stopping it's model. This stopps all the models after logout.

Change-Id: If6cf48f34b426cc86d23891633c5aef40acaade6