core: Add upgrade script for disk alias
Adding an upgrade script for nitialization of disk alias name.
Each disk will be generated with disk alias which will be combined with
the vm name it is associated with and its internal drive mappping.

This change is part of the upgrade process for changing disk to be
entity object with name and description.

Change-Id: I65b411896cba2803ed9009e873441349093478a3