restapi: Snapshot Refactoring
Snapshots were refactored according recent Backend changes.
A Snapshot now represents a point-in-time of a VM.
When the client GETs a snapshot, he will see the snapshot's
meta-data along with the meta-data of the VM in the
point-in-time when the snapshot was taken.

The client will be given links to VM disks, nics and cd-roms.
Also, implementation of restore-snapshot changed, but the client
does it the same way as always.

Since all implementation changed, all existing snapshots tests
became irrelevant. They need to be replaced with new tests.
In this patch I have replaced the tests in BackendSnapshotsResourceTest
(the snapshot collection resource) but I have not replaced the tests in
BackendSnapshotResourceTest (the single snapshot level). Those tests
have been commented out.

In subsequent patches that I will send, I will restore the
tests in BackendSnapshotResourceTest and add some new tests.

Change-Id: Idca90053f20c77b51698fedaca23b339a85920c8