core: Duplicate key violation during add vm device (#806640)
https://bugzilla.redhat.com/806640

The snapshot configuration contains illegal image.
As a result, a new Disk is added to the clone VM.
The Vm device is added as well in that stage.
After this, there is an attemp to copy the vm devices from
the vm in the configuration, but it also includes the image.
As a result , a duplicat key violation occurs.
The solution is to create the new disk without the vm device
at the stage of illegal image handling, and let the vm device
copying create the vm device for it.

Change-Id: Ia6ecca2836456217942aaeacc3fa7b61d56af1f9
Acked-By: