core: Fixing missing VmDevice (#bz806640)
Previous fix to this bug did not create a vm device entry
for the illegal image.

This fix takes care of this issue -
The illegal disk creation does add vm device, but do not add
the image to newImages list. As a result CopyVmDevices will not
create a new vm device for it.

Change-Id: I19e88222cf666321258b49f1f4f239594540c653
Acked-By: