DO NOT SUBMIT core: Fix admin login (#803380)
https://bugzilla.redhat.com/803380

This patch fixes a scenario in which a group
was granted super user role on system, but one of its users
was not added to the system.
The user will fail to login, as it does not exist in DB
and its Id cannot be used for MLA information extracting from DB.
The solutuion shifts the DB addition after the authentication part,
as was prior to the fix of https://bugzilla.redhat.com/784014
("user already logged in" error when trying to log in)
But still maintains seperation from setting the user on the session
(So in LoginAdminUserCommand it will be possible to set this after the
call to MLA).

Change-Id: I133b9cb148dfad2fa678864046a5ee34c37f3895
Acked-By: