core: Fix admin login (#803380)
https://bugzilla.redhat.com/803380

This patch fixes a scenario in which a group
was granted super user role on system, but one of its users
was not added to the system.
The user will fail to login, as it does not exist in DB
and its Id cannot be used for MLA information extracting from DB.
The solutuion shifts the DB addition after the authentication part,
as was prior to the fix of https://bugzilla.redhat.com/784014
("user already logged in" error when trying to log in)
But still maintains seperation from setting the user on the session
(So in LoginAdminUserCommand it will be possible to set this after the
call to MLA).

This was resubmitted following a revert to
http://gerrit.ovirt.org/#change,3114

Change-Id: I3a1bcf611e5492a45d822bf5a55b4ebac8629392
Acked-By: