core: No empty quotas in storage global default
This patch is intended to stop retuning empty quotas in
getQuotaStorageByQuotaGuidWithGeneralDefault.

A boolean parameter was added to getQuotaStorageByStorageGuid to
determine whether empty quotas should or should not be returned, and
getQuotaStorageByQuotaGuidWithGeneralDefault was changed to call it with
false.

In order to test this behaviour, an additional empty quota was added
to fixtures.xml, and the test on the total number of quotas were fixed
accordingly.

Change-Id: Ib2d3d992115b8ba88072fff6e14a54fbc43bcb9e