rest: session-management for REST API
This patch contains code changes for supporting session management in
the REST API.

Clients should work with it as follows:
1. On first request, pass a special Header: "Prefer: persistent-auth",
to tell the server they expect the session to be persistent. The
credentials should also be passed.
2. The server replies with a JSESSIONID cookie.
3. This cookie should be passes back to the server. The prefer header
from step "1" should be passed as well. No credentials needed.
4. Repeat step 3 according to the amount of requests.
5. The client should pass the cookie without the header in case it
wishes to close the session.
6. If the client, for some reason (like session timeout, if the session
was idle for a long time) gets a challenge response from the server then he needs to pass the credentials
again.

See also: http://www.ovirt.org/wiki/Features/RESTSessionManagement

Change-Id: I19bf58daab2d75c2677d90b09112a1876f5e870e