core : Fixing ExportVmTemplateCommand for multiple storage domains
The following patch will fix a retrieve of source storage domains for disks of template
at case that they located at different storages.
Actually the following patch is replacing an old logic which assumed that disk is located on the
same storage.
The patch contains a clean up of uneeded code, duplicate checks and other "strange" stuff

Change-Id: I89dd89adc7f925e5a5e1f5e972ea6da81aa6bea7