core: Add queries for fetching Permitted Storage Domains
The patch adds two queries for fetching Storage Domains:

GetPermittedStorageDomainsByStoragePoolId(StoragePoolID, ActionGroup) -
returns all storage domains within the specified storage-pool-id that
the current user has permissions on with a role that contains the
specified action-group.

GetPermittedStorageDomainsByTemplateId(TemplateID, ActionGroup) -
returns all storage domains on which the specified template resides that
the current user has permissions on with a role that contains the
specified action-group.

The patch refactors GetPermittedStorageDomainsByTemplateIdQuery so it
could be easily inherited by the new query.

Change-Id: I155cb9f187b951b399e99b54007cf24e3163251a