core: Introducing new App errors for Posix-fs validation
Introducing two new error messages -
a. Error that will be reported when the storage server connection
has wrong parameters for the storage type (i.e - if vfsType is
sent for NFS, and not for Posix)
b. Error for indicating that there is an attempt to create
PosixFS Data center for a wrong compatilibity version

Change-Id: I728f394ea537b6eeb4d67229d5de7a6e954c35f5
Acked-By: