core: Add new business entity: Disk
In contrast to BaseDisk, this entity will also hold the storage specific
implementation details, and is in fact a type of BaseDisk. The storage
specific types will inherit from this type and extend it with storage
specific data.

Clients should update their code to use this class instead of DiskImage
directly, to support other possible storage types.

Change-Id: I8abd008b699c3b8e5e1222979708dacb686fce69