core: ValidateSession servlet fixes
Previous commit added this servlet, that validates the session, and
returns the logged-in user in this servlet.
This patch returns successful result only upon session existance + the
logged in user is an administrator user. For now it is not exposed, so
returning always true.

Also, this patch prints the UPN and not the userName, to get the format
"user@domain" instead of just "user".

Change-Id: I61253ad591bd339f1b32ba4129fcbd9258dc811b