core: Made GetVmCustomPropertiesQuery a user query
The query was redefined as a User query, so it could be used in the User
Portal.

Note that the query is a wrapper around getting two Config values which
are defined as user-config values, so it is also safe to consider this
query as a user query.
No additional filtering is needed.

Change-Id: I475d1f4018b187f46e83505dbb6143ad52f0eba4