core: Remove VmPoolDAO.getAllVmPoolsForAdGroup
Removed VmPoolDAO.getAllVmPoolsForAdGroup:
On the one hand, it's based on a stored procedure that doesn't (and
can't) work.
On the other hand, it's only used by GetVmPoolsAttachedToAdGroupQuery,
which isn't used anywhere on it's own right.

This patch removes this entire stack of dead code - the stored
procedure, the DAO, and the query that uses it.

Change-Id: Ib81f8a76d09de32dda62b75e01e70e536ab8df61