core: Refactor Default DAO to receive entity name
Stored procedure names are pretty much standard, and based on the entity
name. The DefaultGenericDaoDbFacade was changed so that it can receive
this name and use default names for the stored procedures.
The implemeting classes now need to call this constructor, and can
optionally override specific procedure names that are not according to
their default format.

Change-Id: I6d5c7706218f81e34f7acf14952782995d0f5046