core: Remove VMDAO.getById
Removed VMDAO.getById - It basically did VMDAO.get, and then added the
network interfaces, which is an abstraction error.

Most of the usages did not need the network interfaces, so the usages
was changes to VMDAO.get.
In the places where the VM object is saved or retruned, an additional
call was added to explicitly add the network interfaces.

Change-Id: I2590f6d85a2bb74129167af1d817d668f4d0fcfb