core: Extract deviceId from specParams of VmDevice
The patch extracts the deviceId from specParams into an explicit
property which will be sent to VDSM, ignored and sent back to engine.

Current VDSM behavior allows to send unused properties which will be
returned back to the engine when sampling VM statuses. It will allows
the engine to identify those devices as originated at engine.

Change-Id: I72c88819c21648b15e9aee5d22890e7d9e2efb31