core: Improved handling of default quota name
The default quota will be called "DefalutQuota-dcXXX".
When updating a previously default quota, the user will be forced to
change its name, and an additional validation is added to both the
UpdateQuota and AddQuota commands so a non-default quota won't have the
name of default quota (i.e., a name that starts with "DefaultQuota-").

Change-Id: I3c58f42a28a21ebe4fdacf09647fe085327b6437