core: remove 03010250 version from schema_version
Adding a custom.sql script under the upgarde/pre_upgrade directory
The purpose of this file is to fix bugs

Currently we have a bug that upgarde 03010250 was cherry-picked and installed after 03010130
This caused all scripts between 03010140 and 03010240 not to run.
This scripts removes the 03010250 from schema_version and updates 03010130 to be the curent version
So, after that , all scripts from 03010140 and up will run
Since 03010250 was a configuration change that is already handled in the new config.sql, no other handling is needed.

Change-Id: If0e9b546ca10fb51f0b7e5a911baa0907281cae1