webadmin: 'phase one' user portal and webadmin localization
https://bugzilla.redhat.com/show_bug.cgi?id=798943
https://bugzilla.redhat.com/show_bug.cgi?id=798942

All the strings in the frontend are separated to 2 groups:
1. localized strings
2. non localize strings

localized strings-
- Localized strings have to be in a Constants or Messages file and not to be
specified directly in the code.
- The Constants and Massages files will have properties files in all the
needed locales. (For example ApplicationConstans.java=>
ApplicationConstants.properties, ApplicationConstants_es.properties...)
- Not all Enums.properties should be localized. The localized enums should
be in LocalizedEnums interface. Just LocalizedEnums.properties file will
be translated. Enums.properties won't be translated.
- The localization of the uiCommonweb resides in compat Constants and
Messages classes.
- The localized fields are- column names, tab names, titles, messages,
buttons text, labels that doesn't contain values.

non localized strings-
- non localized strings are strings in the code that the programmer decided
not to localize (like command names, property names, html tags...)
- All the non localized strings should be marked with a comment in the
format //$NON-NLS-N$ (N represents the number of the string in the line).
- In the next commit I will add to the check style plug-in a check that all
the strings ("*") in the *.java classes have //$NON-NLS-N$ comment.
- The *.ui.xml files can also contain strings.
I localized the *.ui.xml files we have. But, for now, String existence in
these files won't be checked by the check style. So, we should put extra
attention to this issue in the code review.
in compile time.

what is not localized yet (or forever)-
1. search
2. values rendering
3. permission roles description
4. Enums.properties
5. Events and alerts

Change-Id: I74c56a3aed35dec0a378efc58a2e44569eff6089