core: Fixed GetAllDisksByVmIdQueryTest
The test silently assumed that a disk could have a snapshot which is the
disk itself, which is an illegal state.

Since change I6bfc9c6a this state became not only (silently) illegal,
but actually caused a StackOverflowError.

This behavior is inconsequential to the test, which dones't really care
what snapshots are assossiated to the disk, as long as there are some.
This patch creates new random snapshots for the disk, instead of just
reusing the same object.

Change-Id: Ib069780c99a04317a02d5017764c15528f931846