core: creating pool when there is not enough space
Adding a new configuration variable : VmPoolMaxSubsequentFailures (defaulted to 3)
This limits the number of subsequent failures in vm creation in a pool to the configured number
After reaching this number, vm creation loop will stop generating a audit log messages determining
the reason of failure and how many vms from requested were created.
Keep in mind that we count only subsequent failures. The reason for that is that we may have a
temporary problem on a certain VM, so if a VM creation is failed we continue the loop and advance
the subsequent failures counter and if a VM creation succeeded we reset this counter.

Change-Id: I76d485c85e609f4cbea7e2fd37c90937e792c00d
Signed-off-by: Eli Mesika <emesika@redhat.com>