core: Login Commands - add admin flag to VdcUser
The login base command will now update the isAdmin property of the
VdcUser, so it can be shared on the seeion instead of calling
MultiLevelAdministrationHandler each time it's needed (e.g., in user
queries).

LoginAdminUserCommand is left in order to fail the canDoAction if the
user is not an admin, but the actual MLA checking is not done there
anymore.

Change-Id: I73d23480d9d47a199dabeaadc3979e4559d463fa
Signed-off-by: Allon Mureinik <amureini@redhat.com>