core : Start using a Disk object at base object for disk of vms instead of DiskImage
The following patch is introducing a use of Disk object instead of DiskImage.
The change contains simple refactoring, some changes at bussiness logic and fixing some issues.
These is only begginning, additional chnages will be done in the future (Did not added them now,
because of size of the patch)

Change-Id: Iec04f6deed073f58f39dc262e3dc9e30a3851e13
Signed-off-by: Michael Kublin <mkublin@redhat.com>