webadmin: Gluster Create Volume dialog changes
Replica count and stripe count in the volume dialog is disabled.
And the user has to go to Add Bricks dialog to change Replica/Stripe count,
because the count and no. of bricks selected by the user is tightly coupled.
As well as the same dialog is re-used for adding the bricks to existing volumes later.
Add Bricks dialog will be shown whenever the user changes volume type in Volume Dialog.

Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Change-Id: I61163b772c97f70ab055583efd54bc8bfbaafe04