core: VM Interfaces are omitted from templates (#808870)
The patch fixes a bugs in which VM Interfaces are being added to a
template created from a VM nor to a VM created from template.

The cause of the bug is modification made to vm_interface_view by patch
http://gerrit.ovirt.org/#change,3456.

Instead on relying on the view to get the interfaces for the template,
they will be sent as parameter for creating vm device for them.

Change-Id: If226d39eecf972b6985ae24a0d85b8215f7d70d1